requirements.txt
pyserial==3.5
fastapi==0.94.1
pydantic==1.10.6
uvicorn==0.21.0
websockets==10.4